updated uninstall.ps1

+++ pwsh/uninstall.ps1
@@ -491,9 +491,9 @@ foreach ($app in $selectedApps) {
             Stop-Process -n WinXCorners -Force
             Stop-Process -n WinLaunch -Force
             Stop-Process -n ssn -Force
-            Remove-Item -Path "$env:LOCALAPPDATA\WinLaunch" -Recurse -Force
             Remove-Item -Path "$env:LOCALAPPDATA\WinXCorners" -Recurse -Force
             Remove-Item -Path "$env:LOCALAPPDATA\WinMac\hotcorners" -Recurse -Force
+            if (-not (Get-Process -Name WinLaunch -ErrorAction SilentlyContinue)) { Remove-Item -Path "$env:LOCALAPPDATA\WinLaunch" -Recurse -Force }
             Invoke-Output { Uninstall-WinGetPackage -name "Simple Sticky Notes" }
             Invoke-Output { Install-WinGetPackage 9NBLGGH4QGHW }
             $regPath1 = "HKCU:\Software\Microsoft\Windows\CurrentVersion\Run"
